refactor(platform-github): 适配器按领域拆分为四个服务
重构 Phase 2b（GitHub）。把单体 GitHubAdapter 拆为四个领域服务（GitHubConnection /
GitHubPullRequestService / GitHubCommentService / GitHubMediaService，各 extends
platform-core 领域基类），共享一份连接上下文（MutableConnectionContext）；GitHubAdapter
退化为薄门面，过渡期委托扁平方法给对应领域服务并暴露 .connection/.pulls/.comments/.media，
消费方迁移前保持可用。

- 连接配置归连接层自管：移除 env 间接层，client（统一连接封装实例）自管 base 归一、
  clone 协议与 web/git host 派生，并提供 getCloneUrl；各服务只持 client。
- 映射就近维护：领域专属映射下沉为各服务私有方法，公共响应类型抽到 types.ts，
  跨领域 mapUser 留在 mappers.ts。

公开方法签名不变、适配器测试全过、desktop typecheck 通过，行为等价。
